Disable bluetooth health feature
In these days there is no vendor to make BT health device
(BT/EDR HDP profile). And the supported device also be
less. Because we can't get new BT health device, it is hard
to verify APIs for developers and testers. So we disable this
feature in the tizen reference target. If new requirement or
usecase for this feature is coming, we will re-enable the feature.

Change-Id: I4678dd15e569560136f407b32b9a6066fba39a63
Signed-off-by: DoHyun Pyun <dh79.pyun@samsung.com>