Disable Bluetooth GATT server feature in TM1
TM1 target has a HW issue in Bluetooth LE. So it is very unstable
that an application uses GATT Server API. By Tizen 5.0 we notify
this HW problem to SE team, and re-test if LE TC reports the fail.
But it will reduce the confusion to disable GATT server feature
in TM1.

Change-Id: I6211b4a39c05ce3ebeac5ad225fe865adac90c78
Signed-off-by: DoHyun Pyun <dh79.pyun@samsung.com>