Disable Bluetooth GATT client feature in TM
TM1 target has a HW issue in Bluetooth LE. So it is very unstable
that an application uses GATT Client API. By Tizen 5.0 we notify
this HW problem to SE team, and re-test if LE TC reports the fail.
But it will reduce the confusion to disable GATT client feature
in TM1.

Change-Id: I171861637f6fca57713ff7ab80bef33ee33b8c7c
Signed-off-by: DoHyun Pyun <dh79.pyun@samsung.com>